requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==41.0.3
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
future==0.18.3
mysqlclient==2.2.0
pycparser==2.21
PyJWT==1.7.1
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
serializers==0.2.4
sqlparse==0.4.4
typing-extensions==4.7.1

